% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddiff}
\alias{ddiff}
\title{Likelihood for the Ratcliff diffusion model}
\usage{
ddiff(rt, ch, alpha, theta, xi, tau, eta = as.numeric(c(0)),
  stheta = as.numeric(c(0)), stau = as.numeric(c(0)),
  sigma = as.numeric(c(1)), eps = 1e-29, ln = 0L, parYes = 1L)
}
\arguments{
\item{rt}{a vector of responses times ( rt > 0 ).}

\item{ch}{a vector of accuracy/choice values ( ch = {0,1} ).}

\item{alpha}{a vector of upper boundaries at which the evidence
accumulation terminations.}

\item{theta}{a vector of proportions determining the starting
point for the evidence accumulation, where the starting point
zeta = alpha*theta ( 0 >= theta >= 1 ).}

\item{xi}{a vector of drift rates, the rate of evidence accumulation
( xi > 0 ).}

\item{tau}{a vector of residual latencies for the non-decision
component ( tau > 0 ).}

\item{eta}{a vector of standard deviations for the trial-to-trial
variability in the drift rate (eta > 0). A value results in no
trial-to-trial variability.}

\item{stheta}{a vector of the ranges for the uniform variability
about the starting point ( 0 <= theta +/- (stheta/2) <= 1). A
value results in no trial-to-trial variability.}

\item{stau}{a vector of the ranges for the uniform variability
about the residual latency ( 0 <= tau - (stau/2) ). A
value results in no trial-to-trial variability.}

\item{sigma}{a vector of within-trial variabilities, the drift
coefficient ( sigma > 0 ).}

\item{eps}{the margin of error for the infinite sums being calculated
( default is 1e-29 ).}

\item{ln}{indicates whether the likelihood (ln = 0 ) or the
log-likelihood ( ln = 1 ) should be calculated ( default is 0 ).}
}
\value{
A vector of likelihood or log-likelihood values.
}
\description{
Calculates the likelihood or log-likelihood (joint or conditional)
for the Wiener diffusion model using the implementation of Navarro
and Fuss (2009).
}
\section{Notes}{

For unequal vector lengths, values are recycled.
}

\section{References}{

Navarro, D. J., & Fuss, I. G. (2009). Fast and accurate calculations
  for first-passage times in Wiener diffusion models. Journal of
  Mathematical Psychology, 53(4), 222-230.
Ratcliff, R., & Tuerlinckx, F. (2002). Estimating parameters of the
  diffusion model: Approaches to dealing with contaminant reaction
  times and parameter variability. Psychonomic Bulletin & Review, 9,
  438-481.
}

