% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dinvgauss}
\alias{dinvgauss}
\title{Density Function for the Wald Distribution}
\usage{
dinvgauss(t, kappa, xi, sigma, ln = 0L)
}
\arguments{
\item{kappa}{a vector of thresholds determining when a decision
terminates (kappa > 0).}

\item{xi}{a vector of drift rates, or rates of evidence accumulation
(xi > 0).}

\item{sigma}{a vector of the within-trial variabilities
(sigma > 0).}

\item{ln}{indicates whether the log-likelihood should be returned,
where 1 = True, 0 = False (the default).}
}
\value{
Returns a numeric vector consisting of the density (or
  log density) for the model.
}
\description{
Calculates the density function for the Wald distribution (i.e.
an inverse gaussian), parameterized based on Brownian
motion.
}
\section{Notes}{

For unequal vector lengths, values are recycled.
}

