% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{dist_curve}
\alias{dist_curve}
\title{Plot Distribution Function Curves}
\usage{
dist_curve(prm, ch = 1, dist = "wr", b = c(0, 2, 100), rt = NULL,
  opt = list(), ...)
}
\arguments{
\item{prm}{a vector of parameters (distribution dependent).}

\item{ch}{the choice to condition on (i.e. 0 or 1).}

\item{dist}{a string indicating the distribution function to
plot. Can be
\describe{
  \item{\code{wr}}{The Wald race model.}
  \item{\code{ig}}{The inverse gaussian distribution.}
}}

\item{b}{a vector giving the starting and end point of the curve,
and the number of points used to approximate the curve.}

\item{rt}{an optional vector of response times to plot the function
over.}

\item{opt}{a list of named options:
\describe{
  \item{\code{jnt}}{If true, the joint distribution is used.}
  \item{\code{draw}}{If true, the curve is drawn.}
  \item{\code{out}}{If true, output is returned.}
  \item{\code{flip}}{If true, the curve is flipped about the
    x-axis.}
}}

\item{...}{additional plotting parameters.}
}
\value{
Forthcoming
}
\description{
Draws a line for a specified distribution function.
}
\examples{
# Forthcoming
}

