% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{dist_plot}
\alias{dist_plot}
\title{Plot Functions of Distributions}
\usage{
dist_plot(prm = NULL, rt = NULL, ch = 1, dist = "wr", ver = "CDF",
  opt = list(), ...)
}
\arguments{
\item{prm}{a vector of distribution-dependent parameters.}

\item{rt}{an optional vector of response times to plot the function
over.}

\item{ch}{the choice to condition on (i.e. 0 or 1).}

\item{dist}{a string indicating the distribution function to
plot. Can be
\describe{
  \item{\code{wr}}{The Wald race model.}
  \item{\code{ig}}{The inverse gaussian distribution.}
  \item{\code{emg}}{The exponentially modified gaussian
    distribution.}
}}

\item{ver}{the type of function for the distribution. Can be
\describe{
  \item{\code{CDF}}{The cumulative density function.}
  \item{\code{PDF}}{The probability density function.}
  \item{\code{QPE}}{The quantile probability points.}
  \item{\code{HF}}{The hazard function.}
}}

\item{opt}{a list of named options:
\describe{
  \item{\code{b}}{The lower and upper boundaries for
    the sequence of times and the number of points to
    use to approximate the curve.}
  \item{\code{prb}}{The quantile probabilities to
    calculate.}
  \item{\code{pts}}{If false, a line is drawn instead of
    a set of points.}
  \item{\code{draw}}{If true, the line or set of points is
    drawn.}
  \item{\code{out}}{If true, output is returned.}
  \item{\code{flip}}{If true, the curve or set of points
    is flipped about the x-axis.}
}}

\item{...}{additional plotting parameters.}
}
\value{
A list with the x and y-axis plotting values.
}
\description{
Adds a lines/set of points to an existing plot for a
desired function (e.g. cumulative density, probability density,
hazard) of a response time distribution.
}
\examples{
# Forthcoming
}

