% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{lines.seqmodels}
\alias{lines.seqmodels}
\title{Draw a Line Based on a 'seqmodels' Object}
\usage{
\method{lines}{seqmodels}(object, ch = 1, weight = NULL, ...)
}
\arguments{
\item{object}{forthcoming.}

\item{ch}{forthcoming.}

\item{weight}{forthcoming.}

\item{...}{forthcoming.}
}
\value{
Forthcoming
}
\description{
Draw a Line Based on a 'seqmodels' Object
}
