% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pdiff}
\alias{pdiff}
\title{Distribution function for the Wiener diffusion model}
\usage{
pdiff(rt, ch, alpha, theta, xi, tau, sigma = as.numeric(c(1)), eps = 1e-29,
  parYes = 1L)
}
\arguments{
\item{rt}{a vector of responses times ( rt > 0 ).}

\item{ch}{a vector of accuracy/choice values ( ch = {0,1} ).}

\item{alpha}{a vector of upper boundaries at which the evidence
accumulation terminations.}

\item{theta}{a vector of proportions determining the starting
point for the evidence accumulation, where the starting point
zeta = alpha*theta ( 0 >= theta >= 1 ).}

\item{xi}{a vector of drift rates, the rate of evidence accumulation
( xi > 0 ).}

\item{tau}{a vector of residual latencies for the non-decision
component ( tau > 0 ).}

\item{sigma}{a vector of within-trial variabilities, the drift
coefficient ( sigma > 0; default is 1 ).}

\item{eps}{the margin of error for the infinite sums being calculated
( default is 1e-29 ).}

\item{joint}{indicates whether the conditional ( joint = 0 ) or the
joint ( joint = 1) CDF should be returned ( default is 0 ).}
}
\value{
A vector of values for the cumulative distribution.
}
\description{
Calculates the distribution function for the Wiener diffusion
model using the implementation of Blurton et al. (2012).
}
\section{Notes}{

For unequal vector lengths, values are recycled.
}

\section{References}{

}

