% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pinvgauss}
\alias{pinvgauss}
\title{Distribution Function for the Wald Distribution}
\usage{
pinvgauss(t, kappa, xi, sigma)
}
\arguments{
\item{kappa}{a vector of thresholds determining when a decision
terminates (kappa > 0).}

\item{xi}{a vector of drift rates, or rates of evidence accumulation
(xi > 0).}

\item{sigma}{a vector of the within-trial variabilities
(sigma > 0).}
}
\value{
Returns a numeric vector giving P(T <= t | kappa, xi, sigma).
}
\description{
Calculates the distribution function for the Wald distribution (i.e.
an inverse gaussian), parameterized based on Brownian
motion.
}
\section{Notes}{

For unequal vector lengths, values are recycled.
}

