% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot.seqmodels}
\alias{plot.seqmodels}
\title{Generate a Blank Plot for a 'seqmodels' Object}
\usage{
\method{plot}{seqmodels}(x, weight = NULL, ...)
}
\arguments{
\item{weight}{forthcoming.}

\item{...}{forthcoming.}

\item{object}{forthcoming.}
}
\value{
Forthcoming
}
\description{
Generate a Blank Plot for a 'seqmodels' Object
}
