% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{points.seqmodels}
\alias{points.seqmodels}
\title{Add a Set of Points Based on a 'seqmodels' Object}
\usage{
\method{points}{seqmodels}(object, ch = 1, weight = NULL, ...)
}
\arguments{
\item{object}{forthcoming.}

\item{ch}{forthcoming.}

\item{weight}{forthcoming.}

\item{...}{forthcoming.}
}
\value{
Forthcoming
}
\description{
Add a Set of Points Based on a 'seqmodels' Object
}
