% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pwaldrace}
\alias{pwaldrace}
\title{Distribution Function for the Wald Race Model}
\usage{
pwaldrace(rt, ch, k1, xi1, tau1, k0, xi0, tau0, s1 = as.numeric(c(1)),
  s0 = as.numeric(c(1)))
}
\arguments{
\item{rt}{a vector of response times (rt > 0).}

\item{ch}{a vector of choices (ch = {0,1}).}

\item{k1}{the threshold determining when a decision terminates for
choices == 1 ( k1 > 0).}

\item{xi1}{the average rate of evidence accumulation within a trial
for choices == 1 (xi1 > 0).}

\item{tau1}{the residual latency for choices == 1 (tau1 >= 0).}

\item{k0}{the threshold determining when a decision terminates for
choices == 0 ( k0 > 0).}

\item{xi0}{the average rate of evidence accumulation within a trial
for choices == 0 (xi0 > 0).}

\item{tau0}{the residual latency for choices == 0 (tau0 >= 0).}

\item{s1}{the within trial variability for choices == 1 (s1 > 0;
default is 1.0).}

\item{s0}{the within trial variability for choices == 0 (s0 > 0;
default is 1.0).}
}
\value{
Returns the value(s) for the joint distribution function for
the two-choice version of the Wald race model.
}
\description{
Calculates the joint distribution function for a two choice version
of the Wald race model.
}
\section{Notes}{

For unequal vector lengths, values are recycled.
}

