% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qemg_scl}
\alias{qemg_scl}
\title{Title}
\usage{
qemg_scl(p, mu, sigma, lambda, mnRT, mxRT, em_stop, err)
}
\arguments{
\item{variable}{description of variable.}
}
\value{
Function's return value(s).
}
\description{
Short description.
}

