% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{quickdist}
\alias{quickdist}
\alias{plot.quickdist}
\alias{lines.quickdist}
\alias{points.quickdist}
\alias{is.quickdist}
\title{Create a 'quickdist' Object for Quick Plotting}
\usage{
quickdist(dist, type = "PDF", prm = NULL, x = NULL, b = NULL)

\method{plot}{quickdist}(x, weight = NULL, ...)

\method{lines}{quickdist}(object, ch = 1, weight = NULL, ...)

\method{points}{quickdist}(object, ch = 1, weight = NULL, ...)

is.quickdist(x)
}
\arguments{
\item{dist}{a character string indicating the type of distribution to
plot.}

\item{type}{the type of function to compute over the set of plotting
values, either the density ('PDF', 'pdf', 'density', 'Density',
'd'), distribution ('CDF', 'cdf', 'Distribution', 'distribution',
'df', 'DF'), quantile ('QF', 'qf', 'QPE', 'qpe',
'quantile', 'quantiles', 'Quantile', 'Quantiles', 'q'), or hazard
function ('HF', 'hf', 'h', 'Hazard', 'hazard').}

\item{prm}{a vector of parameters. If null, default values are used.
A named vector can be passed in to control the order in which
parameter inputs are extracted.}

\item{x}{an optional vector of response times or probabilities to
plot over.}

\item{b}{an optional vector giving the lower and upper limits as
well as the number of increments used to generate a sequence of
plotting values.}
}
\value{
An object of class 'quickdist', a list consisting of...
\itemize{
  \item \code{d_pv} - x and y-axis values for the density
    function;
  \item \code{p_pv} - x and y-axis values for the distribution
    function;
  \item \code{q_pv} - x and y-axis values for the quantile
    function;
  \item \code{h_pv} - x and y-axis values for the hazard
    function;
  \item \code{type} - a numeric code indicating the
    type of function to plot.
  \item \code{prm} - a named vector with the parameter
    values for the specified distribution.
  \item \code{dist} - A character string with the
    abbreviated label for the specified distribution.
  \item \code{ch} - A numeric code indicating the
    choice for the conditional distribution (for
    two-choice models).
}
}
\description{
A function to create a \code{quickdist} object with methods for the
\code{plot}, \code{lines}, and \code{points} functions,
allowing for quick plotting of density, distribution, quantile,
and hazard functions over a variety of distributions.
}
\details{
Currently supported distributions include...
\itemize{
  \item Exponentially modified Gaussian ('emg', 'EMG');
  \item Shifted inverse Gaussian, or Wald
    ('sig', 'SIG', 'wald', 'Wald', 'ig', 'IG' );
  \item Two-boundary Wiener process, or Wiener
    diffusion ('wp', 'WP', 'wd', 'WD', 'wiener', 'Wiener');
  \item Gaussian, or Normal ('n', 'N', 'gauss', 'Gauss',
    'normal', 'Normal', 'gaussian', 'Gaussian' );
  \item Gamma ('ga', 'gamma', 'Gamma');
  \item Weibull ('wei', 'weibull', 'Weibull');
  \item Log-normal ('ln', 'LN', 'log-normal', 'Log-normal');
  \item Beta ('b', 'B', 'beta', 'Beta');
  \item Exponential ('exp', 'Exp', 'exponential',
    'Exponential');
}

Once a \code{quickdist} object has been created, a blank plot
can be generated via the \code{plot} command. The figure can
then be updated via the \code{lines} or \code{points} functions,
making it easy to add multiple functions to a single figure.
}
\examples{
# Plot density function for shifted inverse Gaussian
obj <- quickdist( 'SIG' )
plot( obj ); lines( obj )
# Add line for density function with new values
# for kappa, xi, and tau (threshold, drift, and
# shift)
obj_2 <- quickdist( 'SIG', prm = c( .5, 2, .3 ) )
lines( obj_2, col = 'red', lty = 2 )

# Plot distribution function for normal distribution
# over custom range and number of points used to
# approximate line
obj <- quickdist( 'Normal', type = 'CDF', b = c(-4,4,10) )
plot( obj ); lines( obj ); points( obj, pch = 19 )

# Plot predicted versus observed quantile functions

# Simulate data for create observed function
prb <- seq( .1, .9, .1 ) # 10\% to 90\% in 10\% increments
# Simulate data from Beta( 5, 2 )
x <- rbeta( 1000, 5, 2 )
q_obs <- list( x = quantile( x, prob = prb ) )
q_obs$y <- prb

# Examine quantile functions for two different predictions
# Predicted under Beta(1,1)
prd_1 <- quickdist( 'b', type = 'QF', x = prb, prm = c( 1, 1 ) )
# Predicted under Beta(5,2)
prd_2 <- quickdist( 'Beta', type = 'QF', x = prb, prm = c( 5, 2 ) )

plot( prd_1 ); points( prd_1, col = 'blue', pch = 19 )
points( prd_2, col = 'red', pch = 15 )
# Observed
points( q_obs$x, q_obs$y, cex = 1.5 )

# Plot hazard function for exponential distribution
# (demonstrates memoryless property)
obj <- quickdist( 'Exponential', type = 'HF' )
plot( obj ); lines( obj )
# Contrast with gamma distribution
obj <- quickdist( 'Gamma', type = 'HF', prm = c( 1.1, 1) )
lines( obj, lty = 2 )

}
