% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qwaldrace}
\alias{qwaldrace}
\title{Inverse Distribution Function for the Wald Race Model}
\usage{
qwaldrace(p, ch, k1, xi1, tau1, k0, xi0, tau0, s1 = as.numeric(c(1)),
  s0 = as.numeric(c(1)), mxRT = 4, em_stop = 20, err = 0.001,
  joint = 1, parYes = 1L)
}
\arguments{
\item{p}{a vector of probabilities ( 0 >= p >= 1).}

\item{ch}{a vector of choices (ch = {0,1}).}

\item{k1}{the threshold determining when a decision terminates for
choices == 1 ( k1 > 0).}

\item{xi1}{the average rate of evidence accumulation within a trial
for choices == 1 (xi1 > 0).}

\item{tau1}{the residual latency for choices == 1 (tau1 >= 0).}

\item{k0}{the threshold determining when a decision terminates for
choices == 0 ( k0 > 0).}

\item{xi0}{the average rate of evidence accumulation within a trial
for choices == 0 (xi0 > 0).}

\item{tau0}{the residual latency for choices == 0 (tau0 >= 0).}

\item{s1}{the within trial variability for choices == 1 (s1 > 0;
default is 1.0).}

\item{s0}{the within trial variability for choices == 0 (s0 > 0;
default is 1.0).}

\item{mxRT}{the maximum RT response time value that the algorithm is applied to.}

\item{err}{the desired degree of precision for the linear interpolation.}

\item{joint}{If 1, indicates that the probabilities are based on the joint
distribution function.}

\item{parYes}{if set to 1, the code is run in parallel.}

\item{em_step}{the maximum number of iterations for the linear interpolation.}
}
\value{
Returns the quantile(s) for the inverse joint distribution function for
the two-choice version of the Wald race model.
}
\description{
Calculates the inverse of the joint distribution function for a two choice
version of the Wald race model using linear interpolation.
}
\section{Notes}{

For unequal vector lengths, values are recycled.
}

