% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qwaldrace_scl}
\alias{qwaldrace_scl}
\title{Title}
\usage{
qwaldrace_scl(prm)
}
\arguments{
\item{variable}{description of variable.}
}
\value{
Function's return value(s).
}
\description{
Short description.
}

