% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remg}
\alias{demg}
\alias{pemg}
\alias{qemg}
\alias{remg}
\title{The Exponentially Modified Gaussian Distribution}
\usage{
remg(N, mu, sigma, lambda)

demg(x, mu, sigma, lambda, ln = 0L)

pemg(x, mu, sigma, lambda)

qemg(p, mu, sigma, lambda, mnRT = -1, mxRT = 4, em_stop = 20L,
  err = 0.001)
}
\arguments{
\item{N}{the number of draws for random generation.}

\item{mu}{a vector of the means for the gaussian variable.}

\item{sigma}{a vector of standard deviations for the gaussian variable
(sigma > 0).}

\item{lambda}{a vector of rates for the exponential variable
(lambda > 0).}

\item{x}{a vector of quantiles.}

\item{ln}{indicates whether the log-likelihood should be returned,
where 1 = True, 0 = False (the default).}
}
\description{
Random generation, density, distribution, and quantile functions for
the distribution of the convolution of gaussian and exponential
random varibles.
}
\section{Details}{

An exponentially modified gaussian distribution describes the sum of
independent normal and exponential random variables, possessing
a characteristic positive skew due to the exponential variable.

For unequal vector lengths, values are recycled.
}

\section{References}{

Forthcoming
}
\examples{
Forthcoming

}

