% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rexwald}
\alias{dexwald}
\alias{rexwald}
\title{The ex-Wald distribution}
\usage{
rexwald(n, kappa, xi, tau)

dexwald(t, kappa, xi, tau, ln = FALSE, ni = FALSE)
}
\arguments{
\item{n}{the number of draws for random generation.}

\item{kappa}{a vector of thresholds determining when a decision
terminates (kappa > 0).}

\item{xi}{a vector of drift rates, or rates of evidence accumulation
(xi \eqn{\ge} 0).}

\item{tau}{a vector of inverse rate parameters (the mean of the
exponentially distributed residual latency; tau > 0).}

\item{t}{a vector of times ( t > 0 ).}

\item{ln}{logical; if \code{TRUE}, probabilities are given as
log(p).}

\item{lower_tail}{logical; if \code{TRUE} (default), probabilities
are \eqn{P(X \le x)} otherwise \eqn{P( X > x)}.}

\item{bounds}{upper limit of the quantiles to explore  for the
approximation via linear interpolation.}

\item{em_stop}{the maximum number of iterations to attempt to
find the quantile via linear interpolation.}

\item{err}{the number of decimals places to approximate the
cumulative probability during estimation of the quantile function.}
}
\value{
\code{dexwald} gives the density, \code{pexwald} gives the
distribution function, \code{qexwald} approximates the quantile
function, and \code{rexwald} generates random deviates.

The length of the result is determined by \code{n} for
\code{rexwald}, and is the maximum of the length of the
numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the
length of the result.
}
\description{
Random generation, density, distribution, and quantile functions
for the convolution of the exponential and Wald distributions,
parameterized for Brownian motion. \code{kappa} refers to the
threshold, \code{xi} refers to the rate of evidence accumulation
towards this threshold, and \code{tau} is the rate parameter
governing the exponentially distribution shift values. Within-trial
variability for the rate of evidence accumulation (the coefficient
of drift) is fixed to 1.
}
\section{Details}{

Forthcoming

A linear interpolation approach is used to approximate the
quantile function, estimating the inverse of the cumulative
distribution function via an iterative procedure. When
the precision of this estimate is set to 8 decimal places,
the approximation will be typically accurate to about half of a
millisecond.
}

\section{References}{

Forthcoming
}
\examples{
Forthcoming

}

