% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rfrechet}
\alias{dfrechet}
\alias{pfrechet}
\alias{qfrechet}
\alias{rfrechet}
\title{The Frechet Distribution}
\usage{
rfrechet(N, alpha, mu)

dfrechet(t, alpha, mu, ln = 0L)

pfrechet(t, alpha, mu)

qfrechet(p, alpha, mu)
}
\arguments{
\item{N}{the number of draws for random generation.}

\item{alpha}{a vector of shape parameters (alpha > 0).}

\item{mu}{a vector of scale parameters (mu > 0).}

\item{t}{a vector of quantiles (typically response times).}

\item{ln}{indicates whether the log-likelihood should be returned,
where 1 = True, 0 = False.}
}
\description{
Random generation, density, distribution, and quantile functions
for the Frechet (inverse Weibull) distribution with a shape and
scale parameter.
}
\section{Details}{

The Frechet distribution is a special case of the generalized extreme
value distribution, the distribution of the maxima of a sequence of
i.i.d. random variables with no lower boundary. The current
parameterization is intended for applications for a variant of the
LBA model (Terry et al., 2015).

For unequal vector lengths, values are recycled.
}

\section{References}{

Terry, A., Marley, A. A. J., Barnwal, E.-J., Wagenmakers, Heathcote,
  A., & Brown, S. D. (2015). Generalising the drift rate distribution
  for linear ballistic accumulators. Journal of Mathematical
  Psychology, 68, 49 - 58.
}
\examples{
# Treatment of illegal values and vectorization
set.seed(100)
rfrechet( 8, c(1,2,-1,1), c(1,.5,1,-1) ) # Returns NA
dfrechet( c(.5, -1), c(1,2,-1,1), c(1,.5,1,-1) ) # Returns 0
pfrechet( c(.5, -1), c(1,2,-1,1),c(1,.4,1,-1) ) # Returns 0

# Distribution function
t = seq(0,2,length=1000)
plot( t, pfrechet(t,3,.5), type = 'l', xlab = 'Time', ylab =
  'Distribution', bty = 'l', yaxt = 'n' )
axis(2,seq(0,1,.5))
# Quantile function
prb = seq( .1, .9, .2 ) # Probabilities
qnt = qfrechet( prb, 3, .5 )
segments( rep(0,length(prb)), prb, qnt, prb )
segments( qnt,rep(0,length(prb)), qnt, prb )

}

