% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvgauss}
\alias{rinvgauss}
\title{Random Deviates from the Wald Distribution}
\usage{
rinvgauss(N, kappa, xi, sigma)
}
\arguments{
\item{N}{the number of draws}

\item{kappa}{a vector of thresholds determining when a decision
terminates (kappa > 0).}

\item{xi}{a vector of drift rates, or rates of evidence accumulation
(xi > 0).}

\item{sigma}{a vector of the within-trial variabilities
(sigma > 0).}
}
\value{
Returns a numeric vector consisting of random draws from
  the Wald distribution.
}
\description{
Generates random draws from a Wald distribution (i.e. an
inverse gaussian), parameterized based on Brownian
motion.
}
\section{Notes}{

For unequal vector lengths, values are recycled.
}

\section{References}{

Michael, J. R., Schucany, W. R., & Haas, R. W. (1976). Generating
  random variates using transformations with multiple roots.
  The American Statistician, 30 (2), 88-90. doi:10.2307/2683801.
}
\examples{
rinvgauss(8,c(1,2,-1,1),c(1,.5,1,-1),c(1,1,1,1))
}

