% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlba}
\alias{rlba}
\alias{dlba}
\alias{plba}
\title{The LBA Model}
\usage{
rlba(
  N,
  A1,
  b1,
  alpha1,
  beta1,
  tau1,
  A0,
  b0,
  alpha0,
  beta0,
  tau0,
  rl = 0L,
  ver = 0L
)

dlba(
  rt,
  ch,
  A1,
  b1,
  alpha1,
  beta1,
  tau1,
  A0,
  b0,
  alpha0,
  beta0,
  tau0,
  rl = 0,
  ln = 0L,
  ver = 0L
)

plba(
  rt,
  ch,
  A1,
  b1,
  alpha1,
  beta1,
  tau1,
  A0,
  b0,
  alpha0,
  beta0,
  tau0,
  rl = 0,
  ver = 0,
  parYes = 1L
)
}
\arguments{
\item{N}{the number of observations to simulate.}

\item{A1}{the upper boundary for the uniformly distributed start
points (A1 >= 0) for choices == 1.}

\item{b1}{the threshold (b1 >= A1) for choices == 1.}

\item{alpha1}{a location/shape parameter for the distribution
of drift rates for choices == 1.}

\item{beta1}{a scale parameter for the distribution of drift
rates for choices == 1.}

\item{A0}{the upper boundary for the uniformly distributed start
points (A0 >= 0) for choices == 0.}

\item{b0}{the threshold (b0 >= A0) for choices == 0.}

\item{alpha0}{a location/shape parameter for the distribution
of drift rates for choices == 0.}

\item{beta0}{a scale parameter for the distribution of drift
rates for choices == 0.}

\item{rl}{if 1, the residual latency impacts the decision rule.}

\item{ver}{indicates which variant of drift rate distribution
for the LBA should be used.}

\item{rt}{a vector of response times (rt > 0).}

\item{ch}{a vector of choices (ch = {0,1}).}

\item{ln}{indicates whether the log-likelihood should be returned,
where 1 = True, 0 = False (the default).}

\item{parYes}{if set to 1, the code is run in parallel.}

\item{mxRT}{the maximum RT response time value that the algorithm is applied to.}

\item{em_step}{the maximum number of iterations for the linear interpolation.}

\item{err}{the desired degree of precision for the linear interpolation.}

\item{joint}{If 1, indicates that the probabilities are based on the joint
distribution function.}
}
\description{
Random generation, density, distribution, and quantile functions
for a two accumulator version of the LBA model (Brown & Heathcote,
2008; Terry et al., 2015).
}
\section{Notes}{

For unequal vector lengths, values are recycled. For random draws,
inadmissible values return NA. For a standard LBA with normally
distributed drifts, the distribution is truncated so that drift
rates will only be positive.
}

\section{References}{

Brown, S. D., & Heathcote, A. (2008). The simplest complete model of
  choice response time: Linear ballistic accumulation. Cognitive
  psychology, 57(3), 153-178.
Terry, A., Marley, A. A. J., Barnwal, E.-J., Wagenmakers, Heathcote,
  A., & Brown, S. D. (2015). Generalising the drift rate distribution
  for linear ballistic accumulators. Journal of Mathematical
  Psychology, 68, 49 - 58.
Singman, H., Brown, S., Gretton, M., Heathcote, A., Voss, A.,
  Voss, J., & Terry, A. (2016). rtdists: Response Time
  Distributions.  R package version 0.6-6.
}

\examples{
Forthcoming

}
