% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlba_1acc}
\alias{plba_1acc}
\alias{rlba_1acc}
\title{A Linear Ballistic Accumulator}
\usage{
rlba_1acc(N, A, b, alpha, beta, ver = 0L)

plba_1acc(t, A, b, alpha, beta, ver = 0L)
}
\arguments{
\item{N}{the number of draws for random generation.}

\item{A}{the upper boundary for the uniformly distributed start
points (A >= 0).}

\item{b}{the threshold for the accumulator (b >= A).}

\item{alpha}{a location/shape parameter for the distribution
of drift rates.}

\item{beta}{a scale parameter for the distribution of drift
rates.}

\item{ver}{indicates which variant of drift rate distribution
for the LBA should be used.}

\item{t}{a vector of quantiles (typically response times).}

\item{ln}{indicates whether the log-likelihood should be returned,
where 1 = True, 0 = False (the default).}
}
\description{
Random generation, density, and distribution functions
for a single linear ballistic accumulator (Brown & Heathcote,
2008; Terry et al., 2015).
}
\note{
For a standard LBA with normally distributed drifts,
the distribution is truncated so that drift rates will only be
positive.
}
\section{References}{

Brown, S. D., & Heathcote, A. (2008). The simplest complete model of
  choice response time: Linear ballistic accumulation. Cognitive
  psychology, 57(3), 153-178.
Terry, A., Marley, A. A. J., Barnwal, E.-J., Wagenmakers, Heathcote,
  A., & Brown, S. D. (2015). Generalising the drift rate distribution
  for linear ballistic accumulators. Journal of Mathematical
  Psychology, 68, 49 - 58.
}
\examples{
# Forthcoming

}

