% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwaldrace}
\alias{rwaldrace}
\title{Random Deviates from the Wald Race Model}
\usage{
rwaldrace(N, k1, xi1, tau1, k0, xi0, tau0, s1 = as.numeric(c(1)),
  s0 = as.numeric(c(1)))
}
\arguments{
\item{N}{the number of observations to simulate.}

\item{k1}{the threshold determining when a decision terminates for
choices == 1 ( k1 > 0).}

\item{xi1}{the average rate of evidence accumulation within a trial
for choices == 1 (xi1 > 0).}

\item{tau1}{the residual latency for choices == 1 (tau1 >= 0).}

\item{k0}{the threshold determining when a decision terminates for
choices == 0 ( k0 > 0).}

\item{xi0}{the average rate of evidence accumulation within a trial
for choices == 0 (xi0 > 0).}

\item{tau0}{the residual latency for choices == 0 (tau0 >= 0).}

\item{s1}{the within trial variability for choices == 1 (s1 > 0).}

\item{s0}{the within trial variability for choices == 0 (s0 > 0).}
}
\value{
Returns a matrix with two columns, the first consiting of
response times, the second indicate the choice (i.e. the winning
accumulator, either 1 or 0).
}
\description{
Simulates a set of response times and choices from a two accumulator
version of the Wald race model (Logan et al., 2014).
}
\section{Notes}{

For unequal vector lengths, values are recycled. Inadmissible values
return NA.
}

