% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwiener_choice}
\alias{rwiener_choice}
\title{Title}
\usage{
rwiener_choice(par)
}
\arguments{
\item{variable}{description of variable.}
}
\value{
Function's return value(s).
}
\description{
Short description.
}

