% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rexwald}
\alias{rexwald}
\alias{dexwald}
\alias{pexwald}
\title{The ex-Wald distribution}
\usage{
rexwald(n, kappa, xi, tau, sigma = as.numeric(c(1)))

dexwald(t, kappa, xi, tau, sigma = as.numeric(c(1)), ln = FALSE, ni = FALSE)

pexwald(
  t,
  kappa,
  xi,
  tau = as.numeric(c(0)),
  sigma = as.numeric(c(1)),
  ln = FALSE,
  lower_tail = TRUE,
  ni = FALSE
)
}
\arguments{
\item{n}{the number of draws for random generation.}

\item{kappa}{a vector of thresholds determining when a decision
terminates (kappa > 0).}

\item{xi}{a vector of drift rates, or rates of evidence accumulation
(xi \eqn{\ge} 0).}

\item{tau}{a vector of inverse rate parameters (the mean of the
exponentially distributed residual latency; tau > 0).}

\item{sigma}{a vector of the within-trial variabilities (the
coefficient of drift; sigma > 0).}

\item{t}{a vector of times ( t > 0 ).}

\item{ln}{logical; if \code{TRUE}, probabilities are given as
log(p).}

\item{lower_tail}{logical; if \code{TRUE} (default), probabilities
are \eqn{P(X \le x)} otherwise \eqn{P( X > x)}.}

\item{bounds}{upper limit of the quantiles to explore  for the
approximation via linear interpolation.}

\item{em_stop}{the maximum number of iterations to attempt to
find the quantile via linear interpolation.}

\item{err}{the number of decimals places to approximate the
cumulative probability during estimation of the quantile function.}
}
\value{
\code{dexwald} gives the density, \code{pexwald} gives the
distribution function, \code{qexwald} approximates the quantile
function, and \code{rexwald} generates random deviates.

The length of the result is determined by \code{n} for
\code{rexwald}, and is the maximum of the length of the
numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycled to the
length of the result.
}
\description{
Random generation, density, distribution, and quantile functions
for the convolution of the exponential and Wald distributions,
parameterized for Brownian motion. \code{kappa} refers to the
threshold, \code{xi} refers to the rate of evidence accumulation
towards this threshold, \code{tau} is the mean for the
exponentially distribution shift values, and \code{sigma} is
the  within-trial variability for the rate of evidence
accumulation (the coefficient of drift, typically fit to 1 for
identification purposes).
}
\section{Details}{

The ex-Wald distribution is the result if one sums two random
variables, one distributed according to the inverse Gaussian
distribution and the other distributed according to an
exponential distribution with rate equal to 1/tau.

The density is therefore computed from the convolution of the
Wald and exponential distributions. Schwarz (2002) provides a
solution for this convolution. If \eqn{\xi^2} is greater than
\eqn{2 \sigma^2 / \tau }, a straightforward analytic solution
exists (Equation 18). Otherwise, the density is the real
part of a complex function. Schwarz provides a solution in
this case using the complex error function (Equation 9 and 22).
To compute the real and imaginary parts of the complex error
function, the density function uses an algorithm from the
Faddeeva package (Johnson, 2012), with an additional wrapper
taken from the S-PLUS script of Heathcote (2004b).

A linear interpolation approach is used to approximate the
quantile function, estimating the inverse of the cumulative
distribution function via an iterative procedure. When
the precision of this estimate is set to 8 decimal places,
the approximation will be typically accurate to about half of a
millisecond.
}

\section{References}{


Heathcote, A. (2004a). Fitting Wald and ex-Wald distributions to
  response time data: An example using functions for the S-PLUS
  package. Behavior Research Methods Instruments & Computers, 36,
  678 - 694.

Heathcote, A. (2004b). rtfit.ssc. Retrieved May 5, 2017 from
  Psychonomic Society Web Archive: http://www.psychonomic.org/ARCHIVE/.

Johnson, S. G. (2012). Faddeeva Package [Computer software].
  Retrieved from
  http://ab-initio.mit.edu/wiki/index.php/Faddeeva_Package#License

Schwarz, W. (2002). On the convolution of the inverse Gaussian and
  the exponential random variables. Communications in Statistics,
  Theory and Methods, 31, 2113 - 2121.
}

\examples{
Forthcoming

}
