% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvgauss}
\alias{dinvgauss}
\alias{pinvgauss}
\alias{qinvgauss}
\alias{rinvgauss}
\title{The Inverse Gaussian Distribution}
\usage{
rinvgauss(N, kappa, xi, sigma)

dinvgauss(t, kappa, xi, sigma, ln = 0L)

pinvgauss(t, kappa, xi, sigma)

qinvgauss(p, kappa, xi, sigma, mxT = 4, em_stop = 20L, err = 1e-04)
}
\arguments{
\item{N}{the number of draws for random generation.}

\item{kappa}{a vector of thresholds determining when a decision
terminates (kappa > 0).}

\item{xi}{a vector of drift rates, or rates of evidence accumulation
(xi > 0).}

\item{sigma}{a vector of the within-trial variabilities
(sigma > 0).}

\item{t}{a vector of quantiles (typically response times).}

\item{ln}{indicates whether the log-likelihood should be returned,
where 1 = True, 0 = False (the default).}

\item{mxT}{the maximum time that the quantile function algorithm is
applied to.}

\item{err}{the desired degree of precision for the linear interpolation
used in the quantile function.}

\item{em_step}{the maximum number of iterations for the linear
interpolation used in the quantile function.}
}
\description{
Random generation, density, distribution, and quantile functions for
the inverse gaussian (or Wald) distribution.
}
\section{Details}{

The inverse gaussian distribution describes the first passage times
through a positive threshold kappa for a space and time homogenous
Wiener diffusion process. The current functions use a parameterization
suitable for sequential sampling models.

For unequal vector lengths, values are recycled.
}

\section{References}{

Michael, J. R., Schucany, W. R., & Haas, R. W. (1976). Generating
  random variates using transformations with multiple roots.
  The American Statistician, 30 (2), 88-90. doi:10.2307/2683801.
}
\examples{
# Treatment of illegal values and vectorization
set.seed(100)
rinvgauss(8,c(1,2,-1,1),c(1,.5,1,-1),c(1,1,1,1)) # Returns NA
dinvgauss( c(.5, -1), c(1,2,-1,1),c(1,.4,1,-1),c(1,1,1,1) ) # Returns 0
pinvgauss( c(.5, -1), c(1,2,-1,1),c(1,.4,1,-1),c(1,1,1,1) ) # Returns 0

# Demonstrates how parameters kappa and xi scale with different values of sigma
t = seq(0,1,length=1000)
plot( t, dinvgauss(t,.1,.2,.1), type = 'l', bty ='l', xlab = 'Time', ylab = 'Density', lwd = 2 )
lines( t, dinvgauss(t,1,2,1), col = 'red', lty = 2, lwd = 2 )

# Distribution function
plot( t, pinvgauss(t,1,2,1), type = 'l', xlab = 'Time', ylab = 'Distribution', bty = 'l', yaxt = 'n' )
axis(2,seq(.2,.8,.2))
# Quantile function
prb = seq( .2, .8, .2 ) # Probabilities
qnt = qinvgauss( prb, 1, 2, 1 )
segments( rep(0,length(prb)), prb, qnt, prb )
segments( qnt,rep(0,length(prb)), qnt, prb )

}

